refactor: use helpers at the top of components and delete hook file
